requirements.txt
pywebview>=5.0
weasyprint>=60.0
openpyxl>=3.1
markdown-it-py>=3.0
python-docx>=1.1
